Revert "Romana rashid"
