+++ .github/workflows/dotnet-core.yml
@@ -25,7 +25,6 @@ jobs:
             6.0.x
             7.0.x
             8.0.x
-            9.0.x
       
       - name: Install dependencies
         run: dotnet restore ${{env.MainSln}}
