+++ source/GridCell/GridCell.csproj
@@ -14,8 +14,6 @@
   </ItemGroup>
   <ItemGroup>
     <PackageReference Include="NumSharp" Version="0.30.0" />
-    <PackageReference Include="OpenCvSharp4" Version="4.10.0.20241108" />
-    <PackageReference Include="OpenCvSharp4.runtime.win" Version="4.10.0.20241108" />
     <PackageReference Include="ScottPlot" Version="4.1.57" />
   </ItemGroup>
   <ItemGroup>
