+++ source/ImageEncoder/HtmImageEncoder.csproj
@@ -25,7 +25,5 @@
   <ItemGroup>
     <PackageReference Include="ImageBinarizer" Version="1.5.6" />
     <PackageReference Include="NeoCortexApi" Version="1.1.2" />
-    <PackageReference Include="OpenCvSharp4" Version="4.10.0.20241108" />
-    <PackageReference Include="OpenCvSharp4.runtime.win" Version="4.10.0.20241108" />
   </ItemGroup>
 </Project>
