--- source/MySEProject/Image_Binarization/Imagebinarizer/Imagebinarizer.sln
@@ -1,28 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.12.35521.163 d17.12
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Imagebinarizer", "Imagebinarizer.csproj", "{CC40DCDA-C765-487D-B1A1-C06A582B889C}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Unittest1", "..\Unittest1\Unittest1.csproj", "{A94BA8ED-9EFF-45E6-9088-0B29113B7133}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{CC40DCDA-C765-487D-B1A1-C06A582B889C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{CC40DCDA-C765-487D-B1A1-C06A582B889C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{CC40DCDA-C765-487D-B1A1-C06A582B889C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{CC40DCDA-C765-487D-B1A1-C06A582B889C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{A94BA8ED-9EFF-45E6-9088-0B29113B7133}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A94BA8ED-9EFF-45E6-9088-0B29113B7133}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A94BA8ED-9EFF-45E6-9088-0B29113B7133}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A94BA8ED-9EFF-45E6-9088-0B29113B7133}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
