--- source/MySEProject/Image_Binarization/Imagebinarizer/ImgBinarize.cs
@@ -1,23 +1,0 @@
-﻿using Daenet.Binarizer.Entities;
-using Daenet.Binarizer;
-using System.IO;
-using System.Linq;
-using System.Collections.Generic;
-
-/// <summary>
-/// Provides methods to binarize images.
-/// </summary>
-public class ImgBinarize
-{
-    /// <summary>
-    /// Binarizes images in a folder and saves the binarized image data to text files.
-    /// </summary>
-    ///<param name="folderPath">The path to the folder containing images to binarize.</param>
-    public static void ImageBinarize()
-    {
-        Console.WriteLine($"Calling function from ImageBinarizer");
-    }
-
-
-
-}
