--- source/MySEProject/Image_Binarization/Unittest1/MSTestSettings.cs
@@ -1,1 +1,0 @@
-﻿[assembly: Parallelize(Scope = ExecutionScope.MethodLevel)]
