--- source/MySEProject/Image_Binarization/Unittest1/Test1.cs
@@ -1,11 +1,0 @@
-﻿namespace Unittest1
-{
-    [TestClass]
-    public sealed class Test1
-    {
-        [TestMethod]
-        public void TestMethod1()
-        {
-        }
-    }
-}
