+++ source/NeoCortexApi.Parallel/NeoCortexApi.Parallel.csproj
@@ -25,8 +25,6 @@
     <PackageReference Include="DotNetActors" Version="1.0.3" />
     <PackageReference Include="ImageBinarizer" Version="1.0.0" />
     <PackageReference Include="Newtonsoft.Json" Version="12.0.2" />
-    <PackageReference Include="OpenCvSharp4" Version="4.10.0.20241108" />
-    <PackageReference Include="OpenCvSharp4.runtime.win" Version="4.10.0.20241108" />
     
   </ItemGroup>
