+++ source/NeoCortexApi/NeoCortexApi.csproj
@@ -78,10 +78,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="DotNetActors" Version="1.0.3" />
     <PackageReference Include="Microsoft.CSharp" Version="4.7.0" />
-    <PackageReference Include="OpenCvSharp4" Version="4.10.0.20241108" />
-    <PackageReference Include="OpenCvSharp4.runtime.win" Version="4.10.0.20241108" />
    
   </ItemGroup>
