--- source/Samples/NeoCortexApiSampleClassifiers/NeoCortexApiSampleClassifiers.csproj
@@ -1,20 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>net8.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="ImageBinarizer" Version="1.6.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\NeoCortexApi\NeoCortexApi.csproj" />
-    <ProjectReference Include="..\..\NeoCortexEntities\NeoCortexEntities.csproj" />
-    <ProjectReference Include="..\..\NeoCortexUtils\NeoCortexUtils.csproj" />
-  </ItemGroup>
-
-</Project>
