--- source/Samples/NeoCortexApiSampleClassifiers/NeoCortexApiSampleClassifiers.sln
@@ -1,22 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.12.35527.113 d17.12
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NeoCortexApiSampleClassifiers", "NeoCortexApiSampleClassifiers.csproj", "{A5688DF9-BA53-419E-B4D9-A2B0FB85108A}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{A5688DF9-BA53-419E-B4D9-A2B0FB85108A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A5688DF9-BA53-419E-B4D9-A2B0FB85108A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A5688DF9-BA53-419E-B4D9-A2B0FB85108A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A5688DF9-BA53-419E-B4D9-A2B0FB85108A}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
