--- source/Samples/NeoCortexApiSampleClassifiers/Program.cs
@@ -1,10 +1,0 @@
-﻿namespace NeoCortexApiSampleClassifiers
-{
-    internal class Program
-    {
-        static void Main(string[] args)
-        {
-            Console.WriteLine("Hello, World!");
-        }
-    }
-}
