+++ source/UnitTestsProject/UnitTestsProject.csproj
@@ -47,8 +47,6 @@
     <PackageReference Include="NeuralNet.MLPerceptron" Version="1.2.2" />
     <PackageReference Include="NeuralNetworks.Core" Version="1.2.1" />
     <PackageReference Include="Microsoft.Extensions.Configuration.CommandLine" Version="8.0.0" />
-    <PackageReference Include="OpenCvSharp4" Version="4.10.0.20241108" />
-    <PackageReference Include="OpenCvSharp4.runtime.win" Version="4.10.0.20241108" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\NeoCortexApi.Parallel\NeoCortexApi.Parallel.csproj" />
