+++ source/WebSocketNeuroVisualizer/WebSocketNeuroVisualizer.csproj
@@ -15,12 +15,6 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="DotNetActors" Version="1.0.3" />
-    <PackageReference Include="OpenCvSharp4" Version="4.10.0.20241108" />
-    <PackageReference Include="OpenCvSharp4.runtime.win" Version="4.10.0.20241108" />
-  </ItemGroup>
-
-  <ItemGroup>
     <ProjectReference Include="..\NeoCortexApi\NeoCortexApi.csproj" />
     <ProjectReference Include="..\NeoCortexEntities\NeoCortexEntities.csproj" />
   </ItemGroup>
